+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Update default JOSM/Vespucci presets from 2.9.0 to 4.1.2
+
 ## 2.0.1 - 2022-02-07
 
 * Fix trunk roads to Urban visualization preset (contribution from @Rikuoja)