+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Improve help how to create map preset on the online help
+* Improve the help about add some QML in queries
 * Fix Python exception if the user input is not recognised as a valid preset
 * Update default JOSM/Vespucci presets from 2.9.0 to 4.1.2